+++ elfo-core/src/stream.rs
@@ -133,11 +133,11 @@ pub struct Stream<M = AnyMessage> {
 #[sealed]
 impl<M: StreamItem> crate::source::SourceHandle for Stream<M> {
     fn is_terminated(&self) -> bool {
-        self.source.lock().is_none()
+        self.source.is_terminated()
     }
 
-    fn terminate(self) {
-        ward!(self.source.lock()).terminate();
+    fn terminate_by_ref(&self) -> bool {
+        self.source.terminate_by_ref()
     }
 }