+++ elfo-core/src/time/delay.rs
@@ -58,11 +58,11 @@ pub struct Delay<M> {
 #[sealed]
 impl<M: Message> crate::source::SourceHandle for Delay<M> {
     fn is_terminated(&self) -> bool {
-        self.source.lock().is_none()
+        self.source.is_terminated()
     }
 
-    fn terminate(self) {
-        ward!(self.source.lock()).terminate();
+    fn terminate_by_ref(&self) -> bool {
+        self.source.terminate_by_ref()
     }
 }