+++ elfo-core/src/time/interval.rs
@@ -61,11 +61,11 @@ pub struct Interval<M> {
 #[sealed]
 impl<M: Message> crate::source::SourceHandle for Interval<M> {
     fn is_terminated(&self) -> bool {
-        self.source.lock().is_none()
+        self.source.is_terminated()
     }
 
-    fn terminate(self) {
-        ward!(self.source.lock()).terminate();
+    fn terminate_by_ref(&self) -> bool {
+        self.source.terminate_by_ref()
     }
 }